[IOTMID-920] Kong related functions moved to a new file
This commit add no new features. Just a rearrange of functions.
